+++ common/src/main/java/com/frostwire/search/FileSearchResult.java
@@ -27,5 +27,5 @@ public interface FileSearchResult extends SearchResult {
 
     String getFilename();
 
-    long getSize();
+    double getSize();
 }
