+++ common/src/main/java/com/frostwire/search/torlock/TorLockSearchResult.java
@@ -39,7 +39,7 @@ public final class TorLockSearchResult extends AbstractTorrentSearchResult {
     private final String detailsUrl;
     private final String torrentUrl;
     private final String infoHash;
-    private final long size;
+    private final double size;
     private final long creationTime;
     private final int seeds;
 
@@ -55,7 +55,7 @@ public final class TorLockSearchResult extends AbstractTorrentSearchResult {
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return size;
     }
