+++ common/src/main/java/com/frostwire/search/zooqle/ZooqleSearchResult.java
@@ -41,7 +41,7 @@ public final class ZooqleSearchResult extends AbstractTorrentSearchResult {
     private final String detailsUrl;
     private final String torrentUrl;
     private final String infoHash;
-    private final long size;
+    private final double size;
     private final long creationTime;
     private final int seeds;
 
@@ -95,7 +95,7 @@ public final class ZooqleSearchResult extends AbstractTorrentSearchResult {
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return size;
     }
